fix: move from saxen to saxes
Saxen issues warnings on encountering malformed data. So for us saxen's warnings
are errors because we do not allow malformed data.

Now, the latest saxen version contains a change to issue a warning when **any**
text appears outside the root element. It issues a warning no matter what the
text is meant to represent. Comments are valid outside the root element, for
instance, but saxen will issue a warning when encountering a comment outside the
root element.

Rather than pursue a fix for saxen, we switch to saxes, which does not do stupid
stuff like this.