fix: chunk the test handling
In prior versions we did not use to chunk the list of children. It is unclear
what changed but in the latest version of Chrome, just passing children to
Process.all results in Chrome raising net::ERR_INSUFFICIENT_RESOURCES. We're
loading about 3600 test files. Apparently, that's too much for Chrome. So now we
batch the children to avoid loading so many at once.

The current solution works but, ideally, a queue that restricts the number of "in
flight" requests would be implemented.